Zamiana float na int

+++ main.py
@@ -5,7 +5,7 @@ def silnia(liczby):
   n_k_tablica_int = [int(x) for x in n_k]
   n = n_k_tablica_int[0]
   k = n_k_tablica_int[1]
-  wynik = mt.factorial(n)/(mt.factorial(k)*mt.factorial(n-k))
+  wynik = int(mt.factorial(n)/(mt.factorial(k)*mt.factorial(n-k)))
   print(wynik)
 
 silnia(input())
